requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
python-dotenv>=1.0.0
Werkzeug==2.3.7
lxml>=4.9.0
webdriver-manager>=4.0.0
Pillow>=10.0.1
pytesseract>=0.3.10